remove redux-form from UserSettings
